requirements.txt
SQLAlchemy
python-dotenv